source/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "color-quantizer-gui"
version = "0.2.0"
description = "A Tkinter desktop app for image color quantization with K-Means."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "BSD-3-Clause" }
authors = [
  { name = "aster" }
]
dependencies = [
  "numpy>=1.26,<2",
  "Pillow>=10,<12",
  "scipy>=1.11,<2",
  "scikit-learn>=1.5,<2",
]
keywords = ["image-processing", "kmeans", "tkinter", "quantization", "desktop-app"]

[project.scripts]
color-quantizer-gui = "color_quantizer_gui.app:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
